requirements.txt
asgiref==3.8.1
Django==4.2.21
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
PyJWT==2.9.0
sqlparse==0.5.3
commitizen==4.8.2
